% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_error_LDA.R
\name{get_error_LDA}
\alias{get_error_LDA}
\title{LDA Related Functions}
\usage{
get_error_LDA(x, y, prior)
}
\arguments{
\item{y}{}
}
\value{

}
\description{
Calculate the error generated from LDA when fitting the nodes.
}
