% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDAtree.R
\name{LDAtree}
\alias{LDAtree}
\title{Main function in LDATree}
\usage{
LDAtree(formula, data, prior = NULL, max_level = 5, min_nsize = NULL)
}
\arguments{
\item{formula}{an object of class 'formula'}

\item{data}{Data frame with response variable in it.}
}
\value{
List
}
\description{
Fit the LDATree, and return the treenodes.
}
