% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_SingleTreee.R
\name{new_SingleTreee}
\alias{new_SingleTreee}
\title{Title}
\usage{
new_SingleTreee(
  x,
  response,
  idxCol,
  idxRow,
  missingMethod,
  splitMethod,
  maxTreeLevel,
  minNodeSize,
  currentLevel = 0,
  parentIndex = 0,
  treeList = structure(list(), class = "SingleTreee")
)
}
\arguments{
\item{x}{The same unimputed raw x from the begining. We change the training size through idxRow.}

\item{treeList}{}
}
\description{
Title
}
