% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.Treee}
\alias{plot.Treee}
\title{Plot a Treee object}
\usage{
\method{plot}{Treee}(object, data, node = -1)
}
\arguments{
\item{object}{a fitted model object of class `Treee`, which is assumed to be
the result of the [Treee()] function.}

\item{data}{Input the original data you used to fit the `Treee` object if you
want the individual plot for each node. Otherwise, you can leave this
parameter blank if you only need the overall tree structure.}

\item{node}{Input the node index that you are interested in. It is set to
`-1` by default, and the overall tree structure is drawn.}
}
\description{
Provide a diagram of the whole tree structure or a scatter plot (or density
plot) for a specific tree node.
}
\section{Overall tree structure}{


  A full tree diagram (via the R package [visNetwork]) is shown if `node` is
  not provided (default is `-1`). The color shows the most common (plurality)
  class inside each node. The size of each terminal node is based on its
  relative sample size. Under every node, you see the plurality class, the
  fraction of the correctly predicted training sample vs. the node's sample
  size, and the node index, respectively. When you click on the node, an
  information panel with more details will appear.
}

\section{Individual plot for each node}{


  The node index and the original training data are required to return a more
  detailed plot within a specific node. The density plot will be provided
  when only two levels are left for the response variable in a node (like in
  a binary classification problem). Samples are projected down to their first
  linear discriminant scores (LD1). A scatter plot will be provided if a node
  contains more than two classes. Samples are projected down to their first
  and second linear discriminant scores.
}

