% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotall.R
\name{plotall}
\alias{plotall}
\title{Plot functions}
\usage{
plotall(fit)
}
\arguments{
\item{fit}{A result from LDAtree function}
}
\value{

}
\description{
Plot the LDA tree (Overall and Single node) using visNetwork and ggplot
}
