% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDATree-package.R
\docType{package}
\name{LDATree-package}
\alias{LDATree}
\alias{LDATree-package}
\title{LDATree: Classification Trees with Linear Discriminant Analysis at Terminal Nodes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A classification tree method that uses LDA (Linear Discriminant Analysis) for variable selection, split determination, and model fitting in terminal nodes. It automatically handles missing values and offers visualization tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Moran79/LDATree}
  \item \url{http://iamwangsiyu.com/LDATree/}
  \item Report bugs at \url{https://github.com/Moran79/LDATree/issues}
}

}
\author{
\strong{Maintainer}: Siyu Wang \email{swang739@wisc.edu} (\href{https://orcid.org/0009-0005-2098-7089}{ORCID}) [copyright holder]

}
\keyword{internal}
