% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.Treee}
\alias{predict.Treee}
\title{Predictions from a fitted Treee object}
\usage{
\method{predict}{Treee}(object, newdata, type = c("response", "prob", "all", "grove"))
}
\arguments{
\item{type}{}
}
\description{
Prediction of test data using LDATree.
}
