% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Treee.R
\name{Treee}
\alias{Treee}
\title{Title}
\usage{
Treee(
  formula,
  data,
  splitMethod = "LDScores",
  pruneMethod = "CV",
  prior = NULL,
  weights = NULL,
  maxTreeLevel = 4,
  minNodeSize = NULL,
  numberOfPrune = 20,
  misClassCost = NULL,
  missingMethod = c("meanFlag", "newLevel"),
  randomSeed = NULL
)
}
\arguments{
\item{randomSeed}{}
}
\description{
Title
}
