% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Treee.R
\name{Treee}
\alias{Treee}
\title{Classification trees with Linear Discriminant Analysis terminal nodes}
\usage{
Treee(
  formula,
  data,
  missingMethod = c("meanFlag", "newLevel"),
  splitMethod = "LDScores",
  pruneMethod = "none",
  numberOfPruning = 10,
  maxTreeLevel = 4,
  minNodeSize = NULL
)
}
\arguments{
\item{formula}{an object of class \code{\link[stats:formula]{stats::formula()}}, which has the form
\code{class ~ x1 + x2 + ...}}

\item{data}{a data frame that contains both predictors and the response.
Missing values are allowed in predictors but not in the response.}

\item{missingMethod}{Missing value solutions for numerical variables and
factor variables. 'mean', 'median', 'meanFlag', 'medianFlag' are available
for numerical variables. 'mode', 'modeFlag', 'newLevel' are available for
factor variables. The word 'Flag' in the methods indicates whether a
missing flag is added or not. The 'newLevel' method means that all missing
values are replaced with a new level rather than imputing them to another
existing value.}

\item{splitMethod}{the splitting rule in LDATree growing process. For now,
'LDScores' is the only available option.}

\item{pruneMethod}{the model selection method in the LDATree growing process,
which controls the size of the tree. By default, it's set to 'none', which
applies a direct stopping rule. Alternatively, 'CV' uses the alpha-pruning
process from CART. Although 'CV' is often more accurate, it can be slower,
especially with large datasets.}

\item{numberOfPruning}{controls the number of cross-validation in the
pruning. It is 10 by default.}

\item{maxTreeLevel}{controls the largest tree size possible for either a
direct-stopping tree or a CV-pruned tree. Adding one extra level (depth)
introduces an additional layer of nodes at the bottom of the current tree.
e.g., when the maximum level is 1 (or 2), the maximum tree size is 3 (or
7).}

\item{minNodeSize}{controls the minimum node size. Think carefully before
changing this value. Setting a large number might result in early stopping
and reduced accuracy. By default, it's set to one plus the number of
classes in the response variable.}
}
\value{
An object of class 'Treee' containing the following components:
\itemize{
\item \code{formula}: the formula passed to the \code{Treee}
\item \code{treee}: A list of all the tree nodes, and each node is an object of class 'TreeeNode'.
\item \code{missingMethod}: the missingMethod passed to the \code{Treee}

An object of class 'TreeeNode' containing the following components:
\item \code{currentIndex}
\item \code{currentLevel}:
\item \code{idxRow}, \code{idxCol}:
\item \code{currentLoss}:
\item \code{accuracy}:
\item \code{proportions}:
\item \code{parent}:
\item \code{children}:
\item \code{misReference}:
\item \code{splitCut}
\item \code{nodeModel}
\item \code{nodePredict}
\item \code{offsprings}
\item \code{offspringLoss}
\item \code{alpha}
}
}
\description{
Fit an LDATree model.
}
\details{
Unlike other classification trees, LDATree integrates LDA throughout the
entire tree-growing process. Here is a breakdown of its distinctive features:
\itemize{
\item The tree searches for the best binary split based on sample quantiles of the first linear discriminant score.
\item An LDA/GSVD model is fitted for each terminal node (For more details, refer to \code{\link[=ldaGSVD]{ldaGSVD()}}).
\item Missing values can be imputed using the mean, median, or mode, with optional missing flags available.
\item By default, the tree employs a direct-stopping rule. However, cross-validation using the alpha-pruning from CART is also provided.
}
}
